+++ mountainsort/src/isosplit5/jisotonic5.cpp
@@ -4,7 +4,7 @@
 #include <stdlib.h>
 #include <algorithm>
 
-void jisotonic5(int N, float* BB, float* MSE, float* AA, float* WW)
+void jisotonic5(bigint N, float* BB, float* MSE, float* AA, float* WW)
 {
     if (N < 1)
         return;
@@ -13,7 +13,7 @@ void jisotonic5(int N, float* BB, float* MSE, float* AA, float* WW)
     float* count = (float*)malloc(sizeof(float) * N);
     float* sum = (float*)malloc(sizeof(float) * N);
     float* sumsqr = (float*)malloc(sizeof(float) * N);
-    int last_index = -1;
+    bigint last_index = -1;
 
     last_index++;
     unweightedcount[last_index] = 1;
@@ -25,7 +25,7 @@ void jisotonic5(int N, float* BB, float* MSE, float* AA, float* WW)
     sumsqr[last_index] = AA[0] * AA[0] * w0;
     MSE[0] = 0;
 
-    for (int j = 1; j < N; j++) {
+    for (bigint j = 1; j < N; j++) {
         last_index++;
         unweightedcount[last_index] = 1;
         w0 = 1;
@@ -55,9 +55,9 @@ void jisotonic5(int N, float* BB, float* MSE, float* AA, float* WW)
         }
     }
 
-    int ii = 0;
-    for (int k = 0; k <= last_index; k++) {
-        for (int cc = 0; cc < unweightedcount[k]; cc++) {
+    bigint ii = 0;
+    for (bigint k = 0; k <= last_index; k++) {
+        for (bigint cc = 0; cc < unweightedcount[k]; cc++) {
             BB[ii + cc] = sum[k] / count[k];
         }
         ii += unweightedcount[k];
@@ -69,7 +69,7 @@ void jisotonic5(int N, float* BB, float* MSE, float* AA, float* WW)
     free(sumsqr);
 }
 
-void jisotonic5_updown(int N, float* out, float* in, float* weights)
+void jisotonic5_updown(bigint N, float* out, float* in, float* weights)
 {
     float* B1 = (float*)malloc(sizeof(float) * N);
     float* MSE1 = (float*)malloc(sizeof(float) * N);
@@ -78,22 +78,22 @@ void jisotonic5_updown(int N, float* out, float* in, float* weights)
     float* in_reversed = (float*)malloc(sizeof(float) * N);
     float* weights_reversed = 0;
 
-    for (int j = 0; j < N; j++) {
+    for (bigint j = 0; j < N; j++) {
         in_reversed[j] = in[N - 1 - j];
     }
     if (weights) {
         weights_reversed = (float*)malloc(sizeof(float) * N);
-        for (int j = 0; j < N; j++) {
+        for (bigint j = 0; j < N; j++) {
             weights_reversed[j] = weights[N - 1 - j];
         }
     }
     jisotonic5(N, B1, MSE1, in, weights);
     jisotonic5(N, B2, MSE2, in_reversed, weights_reversed);
-    for (int j = 0; j < N; j++)
+    for (bigint j = 0; j < N; j++)
         MSE1[j] += MSE2[N - 1 - j];
     float bestval = MSE1[0];
-    int best_ind = 0;
-    for (int j = 0; j < N; j++) {
+    bigint best_ind = 0;
+    for (bigint j = 0; j < N; j++) {
         if (MSE1[j] < bestval) {
             bestval = MSE1[j];
             best_ind = j;
@@ -101,9 +101,9 @@ void jisotonic5_updown(int N, float* out, float* in, float* weights)
     }
     jisotonic5(best_ind + 1, B1, MSE1, in, weights);
     jisotonic5(N - best_ind, B2, MSE2, in_reversed, weights_reversed);
-    for (int j = 0; j <= best_ind; j++)
+    for (bigint j = 0; j <= best_ind; j++)
         out[j] = B1[j];
-    for (int j = 0; j < N - best_ind - 1; j++)
+    for (bigint j = 0; j < N - best_ind - 1; j++)
         out[N - 1 - j] = B2[j];
 
     free(B1);
@@ -115,26 +115,26 @@ void jisotonic5_updown(int N, float* out, float* in, float* weights)
         free(weights_reversed);
 }
 
-void jisotonic5_downup(int N, float* out, float* in, float* weights)
+void jisotonic5_downup(bigint N, float* out, float* in, float* weights)
 {
     float* in_neg = (float*)malloc(sizeof(float) * N);
 
-    for (int j = 0; j < N; j++)
+    for (bigint j = 0; j < N; j++)
         in_neg[j] = -in[j];
     jisotonic5_updown(N, out, in_neg, weights);
-    for (int j = 0; j < N; j++)
+    for (bigint j = 0; j < N; j++)
         out[j] = -out[j];
 
     free(in_neg);
 }
 
-void jisotonic5_sort(int N, float* out, const float* in)
+void jisotonic5_sort(bigint N, float* out, const float* in)
 {
     std::copy(in, in + N, out);
     std::sort(out, out + N);
 
     //	QVector<float> in0(N);
-    //	for (int j=0; j<N; j++) in0[j]=in[j];
+    //	for (bigint j=0; j<N; j++) in0[j]=in[j];
     //	qSort(in0);
-    //	for (int j=0; j<N; j++) out[j]=in0[j];
+    //	for (bigint j=0; j<N; j++) out[j]=in0[j];
 }
